+++ modules/perception/traffic_light/recognizer/classify.cc
@@ -14,8 +14,10 @@
  * limitations under the License.
  *****************************************************************************/
 #include "modules/perception/traffic_light/recognizer/classify.h"
+
 #include <caffe/data_transformer.hpp>
 #include <vector>
+
 #include "modules/common/log.h"
 #include "modules/perception/traffic_light/base/utils.h"
 
@@ -105,7 +107,7 @@ void ClassifyBySimple::ProbToColor(const float *out_put_data, float threshold,
   light->status.confidence = out_put_data[max_color_id];
   AINFO << "Light status recognized as " << name_map[max_color_id];
   AINFO << "Color Prob:";
-  for (int j = 0; j < status_map.size(); j++) {
+  for (size_t j = 0; j < status_map.size(); j++) {
     AINFO << out_put_data[j];
   }
 }
